refactor: saving topostats.__release__ in .topostats file version
Switching to using the internal TopoStats `__release__` for versioning files.

Our extraction of the main release from `__version__` was wrong, that has been fixed so that `__version__` holds the full semantic version as reported by `setuptools_scm`. At the same time we introduce the `__release__` which will always hold the most recent tagged release. This is now going to be used in the `topostats_file_version` field of the HDF5 `.topostats` file. This has the advantage of keeping that version as metadata with the raw and processed data, making it easier to (potentially!) reproduce the results in future (hello FAIR for Research Software Principles).