tests: run tests in parallel via pyproject.toml
Closes #1153

Runs tests in parallel and sets the number of jobs to be `auto` which is calculated as the number of _physical_ cores the CPU has.

Other options include using `logical` which, as the name suggests, is the number of logical CPU cores.

For example the `i9=9980HK` in my laptop has 8 physical cores but each one provides two logical CPU cores, i.e. 16 in total.

Since I don't want to tie up my, or anyone elses, system running the test suite I've opted to set the default to the lower of the two (via `auto`).

This should help speed up some of the tests run in CI too where the number of cores is unknown.