packaging: Upgrade minimum numpy and numpyencoder versions
With updates to [topoly-1.1.0]() newer versions of Numpy are supported. However TensorFlow still requires <2.1.0.

At the same time I discovered there are problems with some of the data types in [numpyencoder](https://github.com/hmallen/numpyencoder).

Last week I reported [Support for np.float_ removed](https://github.com/hmallen/numpyencoder/issues/5) only to find that this had already been addressed in the `master` branch but that the tag applied `v0.30. - NumPy 2.0 Compatibility` whilst building a tar-ball on GitHub hadn't triggered or resulted in a version bump on PyPI.

Today I discovered that `np.complex_` is also deprecated in Numpy 2.0 and whilst I have made a [pull
request](https://github.com/hmallen/numpyencoder/pull/6) to have that addressed in `numpyencoder`, rather than waiting on that I have forked the repository to the AFM-SPM organisation and fixed it there, setting TopoStats to install from the `master` branch of that fork.

If/when a new release of `numpyencoder` is made I will revert to using the PyPI version.