feature: write YAML configuration files from other packages
Closes #1220 (and possibly others but I can't find them at the moment!)

TopoStats modular design, which is being improved in current refactoring, means that it should be easy to extend the
analysis pipelines by developing other packages such as [AFMSlicer](https://github.com/AFM-SPM/AFMSlicer) where work is
under way.

One of the things that will be important is to allow developers of such packages, and in turn users, to generate sample
configuration files which they can change as they desire.

Rather than have the same code duplicated across packages we can use the `io.write_config_with_comments()` function from
TopoStats to load a `<pkg_name>/default_config.yaml` from a package and write that to disk which is what this Pull
Request achieves.

I've included an early version of `docs/advanced/extending.md` to document how to develop extension packages, it _will_
change dramatically as this takes shape as this is new territory for me, but felt it important to document what I'm
doing now so that I can expand and improve on it as things change and lesson are learnt.

**NB** This branch will deliberately target `ns-rse/1102-switching-to-TopoStats-class` as that will be the basis on
which other packages are built.