fix: no commit in released versions to report
Closes #1246

When instantiating the module we extract the release and commit to report in the logs and YAML file that are reported.

This was done by splitting the string, development versions will have the from `#.#.#dev###+g<hash>` and released
versions do not contain the `dev###+g<hash>` component to extract.

We now switch to using the [packaging.version](https://packaging.pypa.io/en/stable/version.html) module to check if we
this is a `dev` (`#.#.#dev###+g<hash>`) or `prerelease` (`#.#.#rc#`) version and set the `TOPOSTATS_BASE_VERSION` and
`TOPOSTATS_COMMIT` conditional on that.